完善应用市场中的详情页以及应用的评分bug的修改

+++ NewCRM/NewCRM.Dto/Dto/TodayRecommendAppDto.cs
@@ -14,7 +14,7 @@ namespace NewCRM.Dto.Dto
 
         public String AppIcon { get; set; }
 
-        public Int32 StartCount { get; set; }
+        public Double StartCount { get; set; }
 
         public Boolean IsInstall { get; set; }
